requirements.txt
# Core AI and Database
openai>=1.3.0
supabase>=2.0.0
tiktoken>=0.5.0
numpy>=1.24.0

# Web Framework
aiohttp>=3.9.0
aiohttp-session>=2.12.0

# Authentication and Security
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Configuration and Environment
python-dotenv>=1.0.0

# File System Monitoring
watchdog>=3.0.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
